Moved multi test case to testbase
